Cleanup and end to end test loading image from post
