+++ app/src/main/java/com/mycompany/traveljournal/datasource/ParseClient.java
@@ -44,17 +44,4 @@ public class ParseClient {
         Parse.initialize(context, "ZFoSsZ6iQBe1CvJaNqio6V0nmlN4V7U4VzboX4J4", "0GDxAZahVe7ibC6pqiMNK6n91fYoh7HRfxXLo5TK");
     }
 
-
-    public static void createPost(byte[] imageBytes, String caption, String description, String userId, double latitude, double longitude) {
-        // Save post w/o image
-
-        // Save image with postId
-
-    }
-
-
-
-
-
-
 }
