+++ app/src/main/java/com/mycompany/traveljournal/detailsscreen/DetailFragment.java
@@ -1,7 +1,5 @@
 package com.mycompany.traveljournal.detailsscreen;
 
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
@@ -9,12 +7,15 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
 
 import com.mycompany.traveljournal.R;
-import com.mycompany.traveljournal.datasource.ImageUploader;
-import com.mycompany.traveljournal.examples.ExampleSavePostToParse;
+import com.mycompany.traveljournal.models.Post;
+import com.parse.FindCallback;
+import com.parse.ParseException;
+import com.squareup.picasso.Picasso;
 
-import java.io.ByteArrayOutputStream;
+import java.util.List;
 
 
 public class DetailFragment extends Fragment {
@@ -27,17 +28,8 @@ public class DetailFragment extends Fragment {
         setUpViews(view);
         setUpListeners();
 
-        Log.wtf(TAG, "1");
-        //ExampleGetPostsFromParse.getPostWithImage();
 
-
-//        //testImage();
-//
-//        ImageView ivPost = (ImageView) view.findViewById(R.id.ivPost);
-//        String url = "http://files.parsetfss.com/0ac0f4de-204e-49bb-8e25-e6937c3c11ae/tfss-0f0fdd9a-585b-47c4-94f9-9a9528be464e-file";
-//        Picasso.with(getActivity()).load(url).into(ivPost);
-
-        ExampleSavePostToParse.uploadPhotoToPost(getActivity());
+        testPostWithImage();
 
 
         return view;
@@ -57,22 +49,33 @@ public class DetailFragment extends Fragment {
         super.onCreate(savedInstanceState);
     }
 
-    private void testImage() {
-        //get byte array
-        Bitmap bmp = BitmapFactory.decodeResource(getResources(), R.drawable.coffee);
-        ByteArrayOutputStream stream = new ByteArrayOutputStream();
-        bmp.compress(Bitmap.CompressFormat.PNG, 100, stream);
-        byte[] byteArray = stream.toByteArray();
+    private void testPostWithImage() {
 
+        String postId = "SZdAAxPZKf";
+        Post.getPostWithId(postId, new FindCallback<Post>() {
+            @Override
+            public void done(List<Post> posts, ParseException e) {
+                if (e == null) {
+                    Post post = posts.get(0);
 
-        ImageUploader uploader = new ImageUploader("1", byteArray);
-        uploader.upload();
+                    ImageView ivPost = (ImageView) getActivity().findViewById(R.id.ivPost);
+                    Picasso.with(getActivity()).load(post.getImageUrl()).into(ivPost);
+
+                } else {
+                    Log.wtf(TAG, "Post not found");
+                }
+            }
+        });
+
 
 
 
 
 
     }
+
+
+
 
 //    private void loadImage() {
 //
