+++ app/src/main/java/com/mycompany/traveljournal/models/Post.java
@@ -1,9 +1,7 @@
 package com.mycompany.traveljournal.models;
 
 import com.parse.FindCallback;
-import com.parse.GetDataCallback;
 import com.parse.ParseClassName;
-import com.parse.ParseFile;
 import com.parse.ParseGeoPoint;
 import com.parse.ParseObject;
 import com.parse.ParseQuery;
@@ -15,7 +13,7 @@ public class Post extends ParseObject{
 
     private String postID; //Parse ids are strings
 
-    private String imageId;
+    private String imageUrl;
     //private ParseFile image;
 
     private String caption;
@@ -64,20 +62,9 @@ public class Post extends ParseObject{
         return getString("trip_id");
     }
 
-    public String getImageId() {
-        return getString("image_id");
+    public String getImageUrl() {
+        return getString("image_url");
     }
-
-    public void doWithPhoto(GetDataCallback callback) {
-        ParseFile image = (ParseFile) this.get("image");
-        image.getDataInBackground(callback);
-    }
-
-//    public void loadPhotoIntoImageView(ImageView) {
-//        ParseFile image = (ParseFile) this.get("image");
-//        image.getDataInBackground(callback);
-//    }
-
 
     public Post() {
 
@@ -153,6 +140,7 @@ public class Post extends ParseObject{
         output += getTripID() + " ";
         output += getLatitude() + " ";
         output += getLongitude()+ " ";
+        output += getImageUrl() + " ";
         return output;
     }
